seoul-ict/project/guardFall/pyproject.toml
[project]
name = "guardfall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe==0.10.21",
    "opencv-python==4.11.0.86",
    "python-telegram-bot==22.0",
    "streamlit==1.27.0",
    "streamlit-autorefresh==1.0.1",
    "streamlit-webrtc==0.45.0",
]
